Criado teste de envio para base e conftest fixtures close #27
