+++ libdevpro/spam/db.py
@@ -1,3 +1,6 @@
+from time import sleep
+
+
 class Sessao:
     contador = 0
     usuarios = []
@@ -18,6 +21,9 @@ class Sessao:
 
 
 class Conexao:
+    def __init__(self):
+        sleep(1)
+
     def gerar_sessao(self):
         return Sessao()
