+++ libdevpro/spam/main.py
@@ -1,0 +1,7 @@
+class EnviadorDeSpam:
+    def __init__(self, sessao, enviador):
+        self.sessao = sessao
+        self.enviador = enviador
+
+    def enviar(self, remetente, assunto, corpo):
+        pass
