+++ libdevpro/tests/test_spam/conftest.py
@@ -1,0 +1,19 @@
+import pytest
+from libdevpro.spam.db import Conexao
+
+
+@pytest.fixture(scope='session')
+def conexao():
+    # Setup
+    conexao_obj = Conexao()
+    yield conexao_obj
+    # Tear Down
+    conexao_obj.fechar()
+
+
+@pytest.fixture
+def sessao(conexao):
+    sessao_obj = conexao.gerar_sessao()
+    yield sessao_obj
+    sessao_obj.roll_back()
+    sessao_obj.fechar()
